requirements.txt
fastapi
supabase
google-genai
python-dotenv
pandas
requests
uvicorn